requirements.txt
anyio==4.2.0
attrs==23.2.0
black==22.12.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.13.1
flake8==5.0.4
fsspec==2024.2.0
h11==0.14.0
huggingface-hub==0.20.3
identify==2.5.33
idna==3.6
iniconfig==2.0.0
isort==5.10.1
Jinja2==3.1.3
loguru==0.6.0
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
mypy==0.982
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.8.0
numpy==1.24.4
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.2.1
py==1.11.0
pycodestyle==2.9.1
pydantic==1.10.11
pyflakes==2.5.0
pytest==7.1.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
tokenizers==0.15.1
tomli==2.0.1
torch==2.0.1
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
virtualenv==20.25.0
